requirements.txt
numpy
pandas
torch
torch_geometric
scikit-learn
pyyaml
geopandas
xarray
geopy